app/requirements.txt
langchain>=0.3.20
langchain-community>=0.3.20
langchain-google-genai>=2.0.11
langchain-huggingface>=0.1.2
chromadb==0.4.22
pypdf==3.17.4
google-generativeai==0.8.3
python-dotenv==1.0.0
streamlit==1.31.0
sentence-transformers==2.7.0
